matrix.csc-class {SparseM}	R Documentation

Class "matrix.csc"

Description

A new class for sparse matrices stored in compressed sparse column format

Objects from the Class

Objects can be created by calls of the form new("matrix.csc", ...).

Slots

ra:

Object of class numeric, a real array of nnz elements containing the non-zero elements of A, stored in column order. Thus, if i<j, all elements of column i precede elements from column j. The order of elements within the column is immaterial.

ja:

Object of class integer, an integer array of nnz elements containing the row indices of the elements stored in ‘ra’.

ia:

Object of class integer, an integer array of n+1 elements containing pointers to the beginning of each column in the arrays ‘ra’ and ‘ja’. Thus ‘ia[i]’ indicates the position in the arrays ‘ra’ and ‘ja’ where the ith column begins. The last, (n+1)st, element of ‘ia’ indicates where the n+1 column would start, if it existed.

dimension:

Object of class integer, dimension of the matrix

Methods

as.matrix.csr

signature(x = "matrix.csc"): ...

as.matrix.ssc

signature(x = "matrix.csc"): ...

as.matrix.ssr

signature(x = "matrix.csc"): ...

as.matrix

signature(x = "matrix.csc"): ...

chol

signature(x = "matrix.csc"): ...

dim

signature(x = "matrix.csc"): ...

t

signature(x = "matrix.csc"): ...

See Also

matrix.csr-class

[Package SparseM version 1.81 Index]