matrix.csc.hb-class {SparseM}	R Documentation

Class "matrix.csc.hb" - Column Compressed Sparse Matrices stored in Harwell-Boeing Format

Description

A class consisting of the coefficient matrix and the right-hand-side of a linear system of equations, initial guess of the solution and the exact solutions if they exist stored in external files using the Harwell-Boeing format.

Objects from the Class

Objects can be created by calls of the form new("matrix.csc.hb", ...).

Slots

ra:

Object of class numeric, ra component of the csc or ssc format of the coefficient matrix, X.

ja:

Object of class integer, ja component of the csc or ssc format of the coefficient matrix, X.

ia:

Object of class numeric, ia component of the csc or ssc format of the coefficient matrix, X.

rhs.ra:

Object of class numeric, ra component of the right-hand-side, y, if stored in csc or ssc format; right-hand-side stored in dense vector or matrix otherwise.

guess:

Object of class numeric or NULL vector of the initial guess of the solutions if they exist; a null vector otherwise.

xexact:

Object of class numeric or NULL vector of the exact solutions, b, if they exist; a null vector otherwise.

dimension:

Object of class integer, dimenson of the coefficient matrix, X.

rhs.dim:

Object of class integer, dimenson of the right-hand-side, y.

rhs.mode:

Object of class character or NULL storage mode of the right-hand-side; can be full storage or same format as the coefficient matrix.

Methods

model.matrix

signature(object = "matrix.csc.hb"): ...

show

signature(object = "matrix.csc.hb"): show() the object, notably also when auto-printing.

See Also

model.matrix, model.response, read.matrix.hb, matrix.ssc.hb-class

[Package SparseM version 1.84-2 Index]