matrix.coo-class {SparseM}	R Documentation

Class "matrix.coo"

Description

A new class for sparse matrices stored in coordinate format

Objects from the Class

Objects can be created by calls of the form new("matrix.coo", ...).

Slots

ra:

Object of class numeric, a real array of nnz elements containing the non-zero elements of A.

ja:

Object of class integer, an integer array of nnz elements containing the column indices of the elements stored in ‘ra’.

ia:

Object of class integer, an integer array of nnz elements containing the row indices of the elements stored in ‘ra’.

dimension:

Object of class integer, dimension of the matrix

Methods

as.matrix.csr

signature(x = "matrix.coo"): ...

as.matrix

signature(x = "matrix.coo"): ...

dim

signature(x = "matrix.coo"): ...

See Also

matrix.csr-class

[Package SparseM version 1.81 Index]