matrix.csr.chol-class {SparseM}	R Documentation

Class "matrix.csr.chol"

Description

A class of objects returned from Ng and Peyton's (1993) block sparse Cholesky algorithm

Objects from the Class

Objects can be created by calls of the form new("matrix.csr.chol", ...).

Slots

nrow:

Object of class integer, number of rows in the linear system of equations

nnzlindx:

Object of class numeric, number of non-zero elements in lindx

nsuper:

Object of class integer, number of supernodes

lindx:

Object of class integer, vector of integer containing, in column major order, the row subscripts of the non-zero entries in the Cholesky factor in a compressed storage format

xlindx:

Object of class integer, vector of integer of pointers for lindx

nnzl:

Object of class numeric, number of non-zero entries, including the diagonal entries, of the Cholesky factor stored in lnz

lnz:

Object of class numeric, contains the entries of the Cholesky factor

log.det:

Object of class numeric, log determinant of the Cholesky factor

xlnz:

Object of class integer, column pointer for the Cholesky factor stored in lnz

invp:

Object of class integer, vector of integer of inverse permutation vector

perm:

Object of class integer, vector of integer of permutation vector

xsuper:

Object of class integer, array containing the supernode partioning

det:

Object of class numeric, determinant of the Cholesky factor

ierr:

Object of class integer, error flag

time:

Object of class numeric execution time

Methods

backsolve

signature(r = "matrix.csr.chol"): ...

as.matrix.csr

signature(x = "matrix.csr.chol", upper.tri=TRUE): ...

See Also

chol, backsolve

[Package SparseM version 1.81 Index]